Replace GenericTestExecutionResult.TestFramework with TestPathSelector
This is a first step towards #35795, by introducing TestPathSelector
which allows a future XML result fixture to properly discard non-leaf
nodes.

Any differences in test paths are instead resolved by adjusting the
selector before passing it into the results fixture, currently only by
specifying a different literal based on the test framework.

This also removes the entire concept of a "pre-normalized" path and
supports escaping characters in the path.